Export protocol from pktm
