+++ lepcapy/include/net_io_ctrl.h
@@ -1,15 +1,18 @@
 #ifndef LEPCAPY_NET_IO_CTRL_H
 #define LEPCAPY_NET_IO_CTRL_H
 
-int thread_net_io(FILE *fp);
-int thread_net_join();
-void *__thread_net_io(void *file_ptr);
-int __thread_net_dequeue(FILE *fp);
+#include "proto_ctrl.h"
+#include "record_queue.h"
+
+int start_net_io(FILE *fp);
+int start_net_join();
+void *__do_net_io(void *file_ptr);
+int __do_net_dequeue(FILE *fp);
 
 /*
  * File I/O Functions
  */
 int __net_io_init(FILE *fp);
-int __net_io_read(FILE *fp, struct queue_node_s*);
+int __net_io_send(FILE *fp, struct queue_node_s*);
 
 #endif //LEPCAPY_NET_IO_CTRL_H
